# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "conTIgo API",
        "description": "Autenticación cookie ACCESS HttpOnly JWT RS256. Obtenga GET /api/v1/auth/csrf y envíe su token en X-XSRF-TOKEN para todo POST, incluso login, refresh y logout. Cookies same-origin; no Bearer header. Paginación base cero, size 1..100, sort: createdAt, code, priority, state. Errores ProblemDetail: 400 validación, 401 autenticación, 403 permiso/CSRF, 404 inexistente, 409 estado/concurrencia, 413 tamaño, 429 intentos.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "cookieAuth": []
        }
    ],
    "paths": {
...
        "/api/v1/tickets": {
            "get": {
                "tags": [
                    "ticket-controller"
                ],
                "operationId": "list",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PENDIENTE",
                                "APROBADO",
                                "ATENDIDO",
                                "RECHAZADO",
                                "CERRADO"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ALTA",
                                "MEDIA",
                                "BAJA"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "createdAt"
                        }
                    },
                    {
                        "name": "ascending",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Results"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto de estado, versión o unicidad",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Falta autenticación válida",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Archivo demasiado grande",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso o CSRF insuficientes",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Formato no compatible con visualización",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ticket-controller"
                ],
                "operationId": "create_1",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "Token de GET /api/v1/auth/csrf, junto con su cookie XSRF-TOKEN.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Create"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketView"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto de estado, versión o unicidad",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Falta autenticación válida",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Archivo demasiado grande",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso o CSRF insuficientes",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Formato no compatible con visualización",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/{id}/reject": {
            "post": {
                "tags": [
                    "ticket-controller"
                ],
                "operationId": "reject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "Token de GET /api/v1/auth/csrf, junto con su cookie XSRF-TOKEN.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Change"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketView"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto de estado, versión o unicidad",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Falta autenticación válida",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Archivo demasiado grande",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso o CSRF insuficientes",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Formato no compatible con visualización",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/{id}/close": {
            "post": {
                "tags": [
                    "ticket-controller"
                ],
                "operationId": "close",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "Token de GET /api/v1/auth/csrf, junto con su cookie XSRF-TOKEN.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Change"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketView"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto de estado, versión o unicidad",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Falta autenticación válida",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Archivo demasiado grande",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso o CSRF insuficientes",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Formato no compatible con visualización",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/{id}/attend": {
            "post": {
                "tags": [
                    "ticket-controller"
                ],
                "operationId": "attend",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "Token de GET /api/v1/auth/csrf, junto con su cookie XSRF-TOKEN.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Change"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketView"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto de estado, versión o unicidad",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Falta autenticación válida",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Archivo demasiado grande",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso o CSRF insuficientes",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Formato no compatible con visualización",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Create": {
                "type": "object",
                "properties": {
                    "dni": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "[0-9]{8}"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 0
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 72,
                        "minLength": 12
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 254,
                        "minLength": 0
                    },
                    "areaId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "position": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 0
                    },
                    "employeeCode": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "EMPLEADO",
                                "JEFE",
                                "GERENTE",
                                "TI"
                            ]
                        },
                        "minItems": 1,
                        "uniqueItems": true
                    },
                    "managedAreas": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    }
                },
                "required": [
                    "areaId",
                    "dni",
                    "email",
                    "name",
                    "password",
                    "position",
                    "roles",
                    "username"
                ]
            },
...
            "Change": {
                "type": "object",
                "properties": {
                    "version": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "observation": {
                        "type": "string",
                        "maxLength": 4000,
                        "minLength": 0
                    }
                },
                "required": [
                    "version"
                ]
            },
...
            "Results": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TicketView"
                        }
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "ProblemDetail": {
                "properties": {
                    "status": {},
                    "detail": {
                        "type": "string"
                    },
                    "instance": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "ACCESS",
                "in": "cookie"
            },
            "refreshCookie": {
                "type": "apiKey",
                "name": "REFRESH",
                "in": "cookie"
            }
        }
    }
}